opslaan

+++ backend/src/main/java/MoneyMachine/models/dtos/requests/DepositRequest.java
@@ -20,7 +20,7 @@ public class DepositRequest  {
     @NotNull
     private String type;
     @NotNull
-    private int initiatedBy;
+    private long initiatedBy;
     @NotNull
     private String message;
