+++ backend/src/main/java/MoneyMachine/models/dtos/requests/TransferRequest.java
@@ -23,7 +23,7 @@ public class TransferRequest  {
     @NotNull
     private String type;
     @NotNull
-    private int initiatedBy;
+    private long initiatedBy;
     @NotNull
     private String message;
